requirements.txt
# Django related requirements
django==1.9.7
django-filter==0.13.0
psycopg2==2.6.1
simplejson==3.8.2
djangorestframework==3.3.3
#django-spillway==0.5.3
git+https://github.com/bkg/django-spillway.git@0104293a75214268436812c20a951717c8a40c40#egg=django_spillway
pylibmc==1.3.0

mapnik==0.1
mapbox-vector-tile==0.2.1
# Must be built prior to GDAL
numpy==1.11.0
GDAL==1.11.2
#greenwich==0.6.1
git+https://github.com/bkg/greenwich.git@ac7b02308d6527e8bd1c174b53e37f55ad59db60#egg=greenwich

Fabric==1.11.1
